bugfix: ensure Custom Attachment uses the correct latest version of the linked Item
